+++ lib/xtool/CMakeLists.txt
@@ -23,7 +23,6 @@ SET(xtool_SRCS
 	iniparser/iniparser.c
 )
 
-
 IF(DXSTREAM_DEBUG)
 	ADD_DEFINITIONS(-DXSTREAM_DEBUG)
 ENDIF(DXSTREAM_DEBUG)
@@ -31,3 +30,5 @@ ENDIF(DXSTREAM_DEBUG)
 INCLUDE_DIRECTORIES(BEFORE .)
 
 ADD_LIBRARY(xtool STATIC ${xtool_SRCS})
+
+TARGET_LINK_LIBRARIES(xtool ${WIN_DBG})
