+++ lib/xtool/xerrhead.cpp
@@ -87,3 +87,158 @@ void XErrorHandler::Exit(void)
 {
 	exit(0);
 }
+
+#ifdef _WIN32
+#include <windows.h>
+#include <dbghelp.h>
+
+#include <stdio.h>
+#include <stdlib.h>
+
+  void windows_print_stacktrace(CONTEXT* context, std::ostringstream &error_string)
+  {
+    SymInitialize(GetCurrentProcess(), 0, true);
+
+    STACKFRAME frame = { 0 };
+
+    /* setup initial stack frame */
+	#ifdef _M_AMD64
+		frame.AddrPC.Offset         = context->Rip;
+		frame.AddrPC.Mode           = AddrModeFlat;
+		frame.AddrStack.Offset      = context->Rsp;
+		frame.AddrStack.Mode        = AddrModeFlat;
+		frame.AddrFrame.Offset      = context->Rbp;
+		frame.AddrFrame.Mode        = AddrModeFlat;
+	#else
+		frame.AddrPC.Offset         = context->Eip;
+		frame.AddrPC.Mode           = AddrModeFlat;
+		frame.AddrStack.Offset      = context->Esp;
+		frame.AddrStack.Mode        = AddrModeFlat;
+		frame.AddrFrame.Offset      = context->Ebp;
+		frame.AddrFrame.Mode        = AddrModeFlat;
+	#endif
+
+    while (StackWalk(
+					#ifdef _M_AMD64
+						IMAGE_FILE_MACHINE_AMD64
+					#else
+						IMAGE_FILE_MACHINE_I386
+					#endif
+					,
+                     GetCurrentProcess(),
+                     GetCurrentThread(),
+                     &frame,
+                     context,
+                     0,
+                     SymFunctionTableAccess,
+                     SymGetModuleBase,
+                     0 ) )
+    {
+		char buffer[sizeof(SYMBOL_INFO) + MAX_SYM_NAME * sizeof(TCHAR)];
+		PSYMBOL_INFO symbol = (PSYMBOL_INFO)buffer;
+		DWORD offset_from_symbol=0;
+    	IMAGEHLP_LINE64 line = {0};
+		symbol->SizeOfStruct = sizeof(SYMBOL_INFO);
+		symbol->MaxNameLen = MAX_SYM_NAME;
+		
+		DWORD64 displacement = 0;
+		HANDLE process = GetCurrentProcess();
+		if (SymFromAddr(process, frame.AddrPC.Offset, &displacement, symbol)) {
+			error_string << "[" << 0 << "] " << symbol->Name;
+			if (SymGetLineFromAddr64( process, frame.AddrPC.Offset, &offset_from_symbol, &line ) )  {
+				error_string << "\t" << line.FileName << "(" << line.LineNumber << ")";
+			}
+			error_string << "\n";
+		} else {
+			error_string << "[" << 0 << "???\n";
+		}
+    }
+
+	ErrH.Abort(error_string.str().c_str());
+
+    SymCleanup( GetCurrentProcess() );
+  }
+
+  LONG WINAPI windows_exception_handler(EXCEPTION_POINTERS * ExceptionInfo)
+  {
+	std::ostringstream error_string;
+    switch(ExceptionInfo->ExceptionRecord->ExceptionCode)
+    {
+      case EXCEPTION_ACCESS_VIOLATION:
+        error_string << "Error: EXCEPTION_ACCESS_VIOLATION\n";
+        break;
+      case EXCEPTION_ARRAY_BOUNDS_EXCEEDED:
+        error_string << "Error: EXCEPTION_ARRAY_BOUNDS_EXCEEDED\n";
+        break;
+      case EXCEPTION_BREAKPOINT:
+        error_string << "Error: EXCEPTION_BREAKPOINT\n";
+        break;
+      case EXCEPTION_DATATYPE_MISALIGNMENT:
+        error_string << "Error: EXCEPTION_DATATYPE_MISALIGNMENT\n";
+        break;
+      case EXCEPTION_FLT_DENORMAL_OPERAND:
+        error_string << "Error: EXCEPTION_FLT_DENORMAL_OPERAND\n";
+        break;
+      case EXCEPTION_FLT_DIVIDE_BY_ZERO:
+        error_string << "Error: EXCEPTION_FLT_DIVIDE_BY_ZERO\n";
+        break;
+      case EXCEPTION_FLT_INEXACT_RESULT:
+        error_string << "Error: EXCEPTION_FLT_INEXACT_RESULT\n";
+        break;
+      case EXCEPTION_FLT_INVALID_OPERATION:
+        error_string << "Error: EXCEPTION_FLT_INVALID_OPERATION\n";
+        break;
+      case EXCEPTION_FLT_OVERFLOW:
+        error_string << "Error: EXCEPTION_FLT_OVERFLOW\n";
+        break;
+      case EXCEPTION_FLT_STACK_CHECK:
+        error_string << "Error: EXCEPTION_FLT_STACK_CHECK\n";
+        break;
+      case EXCEPTION_FLT_UNDERFLOW:
+        error_string << "Error: EXCEPTION_FLT_UNDERFLOW\n";
+        break;
+      case EXCEPTION_ILLEGAL_INSTRUCTION:
+        error_string << "Error: EXCEPTION_ILLEGAL_INSTRUCTION\n";
+        break;
+      case EXCEPTION_IN_PAGE_ERROR:
+        error_string << "Error: EXCEPTION_IN_PAGE_ERROR\n";
+        break;
+      case EXCEPTION_INT_DIVIDE_BY_ZERO:
+        error_string << "Error: EXCEPTION_INT_DIVIDE_BY_ZERO\n";
+        break;
+      case EXCEPTION_INT_OVERFLOW:
+        error_string << "Error: EXCEPTION_INT_OVERFLOW\n";
+        break;
+      case EXCEPTION_INVALID_DISPOSITION:
+        error_string << "Error: EXCEPTION_INVALID_DISPOSITION\n";
+        break;
+      case EXCEPTION_NONCONTINUABLE_EXCEPTION:
+        error_string << "Error: EXCEPTION_NONCONTINUABLE_EXCEPTION\n";
+        break;
+      case EXCEPTION_PRIV_INSTRUCTION:
+        error_string << "Error: EXCEPTION_PRIV_INSTRUCTION\n";
+        break;
+      case EXCEPTION_SINGLE_STEP:
+        error_string << "Error: EXCEPTION_SINGLE_STEP\n";
+        break;
+      case EXCEPTION_STACK_OVERFLOW:
+        error_string << "Error: EXCEPTION_STACK_OVERFLOW\n";
+        break;
+      default:
+        error_string << "Error: Unrecognized Exception\n";
+        break;
+    }
+    /* If this is a stack overflow then we can't walk the stack, so just show
+      where the error happened */
+    windows_print_stacktrace(ExceptionInfo->ContextRecord, error_string);
+
+    return EXCEPTION_EXECUTE_HANDLER;
+  }
+
+  void set_signal_handler()
+  {
+    SetUnhandledExceptionFilter(windows_exception_handler);
+  }
+#else
+	void set_signal_handler();
+#endif
