+++ src/road.cpp
@@ -73,156 +73,6 @@
 #include "palette.h"
 #include "sound/hsound.h"
 
-#ifdef _WIN32
-#include <windows.h>
-#include <dbghelp.h>
-
-#include <stdio.h>
-#include <stdlib.h>
-
-  void windows_print_stacktrace(CONTEXT* context)
-  {
-    SymInitialize(GetCurrentProcess(), 0, true);
-
-    STACKFRAME frame = { 0 };
-
-    /* setup initial stack frame */
-	#ifdef _M_AMD64
-		frame.AddrPC.Offset         = context->Rip;
-		frame.AddrPC.Mode           = AddrModeFlat;
-		frame.AddrStack.Offset      = context->Rsp;
-		frame.AddrStack.Mode        = AddrModeFlat;
-		frame.AddrFrame.Offset      = context->Rbp;
-		frame.AddrFrame.Mode        = AddrModeFlat;
-	#else
-		frame.AddrPC.Offset         = context->Eip;
-		frame.AddrPC.Mode           = AddrModeFlat;
-		frame.AddrStack.Offset      = context->Esp;
-		frame.AddrStack.Mode        = AddrModeFlat;
-		frame.AddrFrame.Offset      = context->Ebp;
-		frame.AddrFrame.Mode        = AddrModeFlat;
-	#endif
-
-    while (StackWalk(
-					#ifdef _M_AMD64
-						IMAGE_FILE_MACHINE_AMD64
-					#else
-						IMAGE_FILE_MACHINE_I386
-					#endif
-					,
-                     GetCurrentProcess(),
-                     GetCurrentThread(),
-                     &frame,
-                     context,
-                     0,
-                     SymFunctionTableAccess,
-                     SymGetModuleBase,
-                     0 ) )
-    {
-		char buffer[sizeof(SYMBOL_INFO) + MAX_SYM_NAME * sizeof(TCHAR)];
-		PSYMBOL_INFO symbol = (PSYMBOL_INFO)buffer;
-		DWORD offset_from_symbol=0;
-    	IMAGEHLP_LINE64 line = {0};
-		symbol->SizeOfStruct = sizeof(SYMBOL_INFO);
-		symbol->MaxNameLen = MAX_SYM_NAME;
-		
-		DWORD64 displacement = 0;
-		HANDLE process = GetCurrentProcess();
-		if (SymFromAddr(process, frame.AddrPC.Offset, &displacement, symbol)) {
-			printf("[%i] %s", 0, symbol->Name);
-			if (SymGetLineFromAddr64( process, frame.AddrPC.Offset, &offset_from_symbol, &line ) )  {
-				std::cout << "\t" << line.FileName << "(" << line.LineNumber << ")\n";
-			}
-		} else {
-			printf("[%i] ???\n", 0);
-		}
-    }
-
-    SymCleanup( GetCurrentProcess() );
-  }
-
-  LONG WINAPI windows_exception_handler(EXCEPTION_POINTERS * ExceptionInfo)
-  {
-    switch(ExceptionInfo->ExceptionRecord->ExceptionCode)
-    {
-      case EXCEPTION_ACCESS_VIOLATION:
-        fputs("Error: EXCEPTION_ACCESS_VIOLATION\n", stderr);
-        break;
-      case EXCEPTION_ARRAY_BOUNDS_EXCEEDED:
-        fputs("Error: EXCEPTION_ARRAY_BOUNDS_EXCEEDED\n", stderr);
-        break;
-      case EXCEPTION_BREAKPOINT:
-        fputs("Error: EXCEPTION_BREAKPOINT\n", stderr);
-        break;
-      case EXCEPTION_DATATYPE_MISALIGNMENT:
-        fputs("Error: EXCEPTION_DATATYPE_MISALIGNMENT\n", stderr);
-        break;
-      case EXCEPTION_FLT_DENORMAL_OPERAND:
-        fputs("Error: EXCEPTION_FLT_DENORMAL_OPERAND\n", stderr);
-        break;
-      case EXCEPTION_FLT_DIVIDE_BY_ZERO:
-        fputs("Error: EXCEPTION_FLT_DIVIDE_BY_ZERO\n", stderr);
-        break;
-      case EXCEPTION_FLT_INEXACT_RESULT:
-        fputs("Error: EXCEPTION_FLT_INEXACT_RESULT\n", stderr);
-        break;
-      case EXCEPTION_FLT_INVALID_OPERATION:
-        fputs("Error: EXCEPTION_FLT_INVALID_OPERATION\n", stderr);
-        break;
-      case EXCEPTION_FLT_OVERFLOW:
-        fputs("Error: EXCEPTION_FLT_OVERFLOW\n", stderr);
-        break;
-      case EXCEPTION_FLT_STACK_CHECK:
-        fputs("Error: EXCEPTION_FLT_STACK_CHECK\n", stderr);
-        break;
-      case EXCEPTION_FLT_UNDERFLOW:
-        fputs("Error: EXCEPTION_FLT_UNDERFLOW\n", stderr);
-        break;
-      case EXCEPTION_ILLEGAL_INSTRUCTION:
-        fputs("Error: EXCEPTION_ILLEGAL_INSTRUCTION\n", stderr);
-        break;
-      case EXCEPTION_IN_PAGE_ERROR:
-        fputs("Error: EXCEPTION_IN_PAGE_ERROR\n", stderr);
-        break;
-      case EXCEPTION_INT_DIVIDE_BY_ZERO:
-        fputs("Error: EXCEPTION_INT_DIVIDE_BY_ZERO\n", stderr);
-        break;
-      case EXCEPTION_INT_OVERFLOW:
-        fputs("Error: EXCEPTION_INT_OVERFLOW\n", stderr);
-        break;
-      case EXCEPTION_INVALID_DISPOSITION:
-        fputs("Error: EXCEPTION_INVALID_DISPOSITION\n", stderr);
-        break;
-      case EXCEPTION_NONCONTINUABLE_EXCEPTION:
-        fputs("Error: EXCEPTION_NONCONTINUABLE_EXCEPTION\n", stderr);
-        break;
-      case EXCEPTION_PRIV_INSTRUCTION:
-        fputs("Error: EXCEPTION_PRIV_INSTRUCTION\n", stderr);
-        break;
-      case EXCEPTION_SINGLE_STEP:
-        fputs("Error: EXCEPTION_SINGLE_STEP\n", stderr);
-        break;
-      case EXCEPTION_STACK_OVERFLOW:
-        fputs("Error: EXCEPTION_STACK_OVERFLOW\n", stderr);
-        break;
-      default:
-        fputs("Error: Unrecognized Exception\n", stderr);
-        break;
-    }
-    fflush(stderr);
-    /* If this is a stack overflow then we can't walk the stack, so just show
-      where the error happened */
-    windows_print_stacktrace(ExceptionInfo->ContextRecord);
-
-    return EXCEPTION_EXECUTE_HANDLER;
-  }
-
-  void set_signal_handler()
-  {
-    SetUnhandledExceptionFilter(windows_exception_handler);
-  }
-#endif
-
 #if defined(__unix__) || defined(__linux__) || defined(__APPLE__)
 #include <locale.h>
 #endif
@@ -2799,8 +2649,6 @@ void LoadingMessage(int flush)
 	char* str = getRC();
 	sysfont.draw(8,XGR_MAXY - 35,(unsigned char*)str,255,-1);
 	XBuffer buf;
-	int *v=0;
-    *v = 0;   // trigger the fault
 	if (lang() == RUSSIAN) {
         buf < Convert(zSTR_LOADING_800_RU);
 	} else if (lang() == GERMAN) {
